pyproject.toml
[project]
name = "wa-trail-guide"
version = "0.1.0"
description = "Weekly dashboard summarizing the top WA hiking trails from WTA trip reports."
requires-python = ">=3.11"
dependencies = [
    "requests>=2.32",
    "beautifulsoup4>=4.12",
    "openai>=1.50",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
]

[tool.uv]
package = false
